Big reorganization so that heavy-weight compile time code doesn't pollute downstream compile times
