+++ CMakeLists.txt
@@ -13,7 +13,7 @@ include_directories(${CMAKE_SOURCE_DIR}/include)
 
 find_library(RE2_LIBRARY re2)
 
-add_library(testlib SHARED test-lib.cc test-lib.hpp include/refl_base.hpp)
+add_library(testlib SHARED test-lib.cc test-lib.hpp include/refl_base.hpp include/casts_table.hpp)
 add_executable(test test.cc)
 target_link_libraries(testlib Boost::filesystem dl ${RE2_LIBRARY})
 target_link_libraries(test testlib)
