+++ test-lib.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <refl_base.hpp>
-#include <boost/dll/runtime_symbol_info.hpp>
 
 struct A {
 	using ReflBases = CxxFFI::DefineBases<>;
@@ -18,12 +17,5 @@ struct C : virtual A {
 struct D : B, C {
 	using ReflBases = CxxFFI::DefineBases<B, C>;
 };
-
-
-namespace CxxFFI {
-	template<> struct APIFilter<A> {
-		using type = boost::mpl::bool_<true>;
-	};
-}
 
 const char * castsTable();
